+++ api/data_dragon.go
@@ -1,4 +1,4 @@
-package golio
+package api
 
 import (
 	"encoding/json"
@@ -19,7 +19,7 @@ const (
 )
 
 var (
-	regionToRealmRegion = map[region]string{
+	regionToRealmRegion = map[Region]string{
 		RegionEuropeWest:        "euw",
 		RegionEuropeNorthEast:   "eun",
 		RegionJapan:             "jp",
@@ -44,7 +44,7 @@ type DataDragonClient struct {
 }
 
 // NewDataDragonClient returns a new client for the Data Dragon service.
-func NewDataDragonClient(client Doer, region region, logger log.FieldLogger) *DataDragonClient {
+func NewDataDragonClient(client Doer, region Region, logger log.FieldLogger) *DataDragonClient {
 	c := &DataDragonClient{
 		client: client,
 		logger: logger.WithField("client", "data dragon"),
