+++ api/riot_api_test.go
@@ -1,4 +1,4 @@
-package golio
+package api
 
 import (
 	"fmt"
